+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.9.1
 
 - Return error if send message request receives a non-200 status
+- Return error if webhook registration/deletion requests receive a non-200 status
 
 ## v0.9.0